Allow use of decorator as a function
Added new tests that validate the decorator usage as a function.

Fixed tests so that they don't have to manually add the parent folder to
the path. By adding an __init__.py file to the tests folder turns it into
a formal Python package and pytest can now figure things out automatically.