Implement no-inheritance message serialization
This is an alternative implementation of the message serialization code
that does not rely in messages inheriting from the WebSocketMessage class.
Instead the id and type are inserted in the class via the decorator
function and then extracted during serialization. The id and type are
checked during deserialization before the messages are passed to the
message factory for dispatching.