Add a simple RPC mechanism on top of WS messages
Bumped version to 0.4.0

New support allow the definition of WebSocket API servers and clients
that can be used to bind messages to methods on the server side and bind
methods to messages on the client side. The API supporting code will use
the message objects as schema validators and will use them to serialize
and deserialize the arguments and the return values on both the client and
server sides.

Also in this change:

- Moved a couple of message constants to const.py
- Refactored message deserialization code into a helper function
- Added a couple of new tests to validate API servers and clients